+++ backends/backend-teavm/src/main/java/com/github/xpenatan/gdx/backends/teavm/config/AssetsCopy.java
@@ -301,9 +301,6 @@ public class AssetsCopy {
     @Deprecated
     public static AssetType getType(String file) {
         String extension = extension(file).toLowerCase();
-        if(isImage(extension)) return AssetType.Image;
-        if(isAudio(extension)) return AssetType.Audio;
-        if(isText(extension)) return AssetType.Text;
         return AssetType.Binary;
     }
 
@@ -312,19 +309,5 @@ public class AssetsCopy {
         int dotIndex = name.lastIndexOf('.');
         if(dotIndex == -1) return "";
         return name.substring(dotIndex + 1);
-    }
-
-    private static boolean isImage(String extension) {
-        return extension.equals("jpg") || extension.equals("jpeg") || extension.equals("png") || extension.equals("bmp") || extension.equals("gif");
-    }
-
-    private static boolean isText(String extension) {
-        return extension.equals("json") || extension.equals("xml") || extension.equals("txt") || extension.equals("glsl")
-                || extension.equals("fnt") || extension.equals("pack") || extension.equals("obj") || extension.equals("atlas")
-                || extension.equals("g3dj");
-    }
-
-    private static boolean isAudio(String extension) {
-        return extension.equals("mp3") || extension.equals("ogg") || extension.equals("wav");
     }
 }