+++ backends/backend-teavm/src/main/java/com/github/xpenatan/gdx/backends/teavm/preloader/AssetType.java
@@ -4,11 +4,11 @@ package com.github.xpenatan.gdx.backends.teavm.preloader;
  * @author xpenatan
  */
 public enum AssetType {
-    Image("i"), Audio("a"), Text("t"), Binary("b"), Directory("d");
+    Binary("b"), Directory("d");
 
     public final String code;
 
-    private AssetType(String code) {
+    AssetType(String code) {
         this.code = code;
     }
 }