+++ CHANGES.md
@@ -5,6 +5,8 @@
 - Call dispose when browser closes
 - Improve clipboard text copy/paste.
 - Change default sound/music api to howler.js
+- add shouldEncodePreference config
+- add localStoragePrefix config
 
 [1.0.0-b9]
 - add TeaClassFilter printAllowedClasses() and printExcludedClasses()